requirements.txt
easydict
polars
pyyaml
torchmetrics
transformers
Pillow
tqdm
icrawler
six
bs4
lxml